Fix synchronize buffer advice on set buffer interface
We advise buffer-changing functions so that changing to a buffer "out
of band" still keeps buffers synchronized with buffer rings e.g. in
terms of recency. For this, we'd prefer to advise low-level interfaces
so that any code that changes the buffer will be intercepted, and not
only the few interfaces we may happen to be aware of, such as
`switch-to-buffer`. Unfortunately, as `set-buffer` is not an ELisp
function, it cannot be advised in this way, so we revert to explicitly
advising known interfaces, for now, preferring low-level ones where
possible.